Add new download profile concern
